% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JRF.R
\name{iJRF_permutation}
\alias{iJRF_permutation}
\title{Joint Random Forest for the simultaneous estimation of multiple related networks}
\usage{
iJRF_permutation(X, W=NULL, ntree=NULL, mtry=NULL,
              genes.name, seed, to.store=NULL)
}
\arguments{
\item{X}{List object containing expression data for each class, \code{X=list(x_1,x_2, ... )} where \code{x_j} is a \code{(p x n_j)} matrix with rows corresponding to genes and columns to samples. Missing values are not allowed.}
\item{W}{\code{(p x p)} Optional symmetric matrix containing sampling scores. When omitted, the standard JRF algorithm without weighted sampling scheme will be implemented. Element \code{(i,j)} contains interaction score \code{(i - j)}. Scores must be non-negative. Larger value of sampling score corresponds to higher likelihood of gene \code{i} interacting with gene \code{j}. Columns and rows of \code{W} must be in the same order as the columns of \code{X}. Sampling scores \code{W} are computed considering one prior data such as protein-protein interactions.}
\item{ntree}{Numeric value: number of trees. If omitted, \code{ntree} is set to \code{1000}}
\item{mtry}{Numeric value: number of predictors to be sampled at each node. If omitted, \code{mtry} is set to the square root of the number of variables.}
\item{genes.name}{Vector containing genes name. The order needs to match the rows of \code{x_j}.}
\item{seed}{Integer. Seed of permutation}

\item{to.store}{Optional Integer. Total number of importance scores to be stored. When omitted, all importance scores will be stored. Note that to compute FDR we do not need all \code{(p-p) x p / 2} importance scores where \code{p} is the total number of proteins/genes, a sufficiently large number would work. This number is usually chosen based on the number of nodes. Suggested value is \code{p x 20}. }
}
\value{
A matrix with \code{I} rows and \code{C + 2} columns where \code{I} is the total number of gene-gene interactions and \code{C} is the number of classes. The first two columns contain gene names for each interaction while the remaining columns contain importance scores for different classes. 
}
\description{
Derive importance score for models of class iJRF based on one permuted data set.
}
\examples{

 # --- Generate data sets
 nclasses=2               # number of data sets / classes
 n1<-n2<-20               # sample size for each data sets
 p<-5                   # number of variables (genes)
 genes.name<-paste("G",seq(1,p),sep="")   # genes name
 W<-abs(matrix(rnorm(p*p),p,p))    # generate weights for relationships

 data1<-matrix(rnorm(p*n1),p,n1)       # generate data1
 data2<-matrix(rnorm(p*n2),p,n1)       # generate data2
 
 # --- Run JRF and obtain importance score of interactions for each class
  out<-iJRF_permutation(X=list(data1,data2),W=W,
                         genes.name=genes.name,seed=1)

}
\references{
Petralia, F., Song, W.M., Tu, Z. and Wang, P. (2016). New method for joint network analysis reveals common and different coexpression patterns among genes and proteins in breast cancer. \emph{Journal of proteome research}, \bold{15}(3), pp.743-754.

A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. \emph{R News} \bold{2}, 18--22.
}
